Fix FTP server view navigation
