+++ DesktopApplicationTemplate.UI/ViewModels/MainViewModel.cs
@@ -123,7 +123,6 @@ namespace DesktopApplicationTemplate.UI.ViewModels
                 "Heartbeat" => (WpfBrushes.LightPink, WpfBrushes.DeepPink),
                 "SCP" => (WpfBrushes.LightCyan, WpfBrushes.CadetBlue),
                 "MQTT" => (WpfBrushes.LightGoldenrodYellow, WpfBrushes.Goldenrod),
-                "FTP" => (WpfBrushes.LightSteelBlue, WpfBrushes.SteelBlue),
                 "FTP Server" => (WpfBrushes.LightSteelBlue, WpfBrushes.SteelBlue),
                 _ => (WpfBrushes.LightGray, WpfBrushes.Gray)
             };
