+++ DesktopApplicationTemplate.UI/Views/FTPServiceView.xaml
@@ -1,64 +1,48 @@
 <Page x:Class="DesktopApplicationTemplate.UI.Views.FTPServiceView"
       xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
       xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-      xmlns:helpers="clr-namespace:DesktopApplicationTemplate.UI.Helpers"
       xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
       xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
       mc:Ignorable="d">
-    <Page.Resources>
-        <helpers:StringNullOrEmptyToVisibilityConverter x:Key="StringNullOrEmptyToVisibilityConverter" />
-    </Page.Resources>
     <Grid Margin="10">
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition Width="2*"/>
-            <ColumnDefinition Width="3*"/>
-        </Grid.ColumnDefinitions>
-        <StackPanel Grid.Column="0" Margin="10">
-            <Grid Margin="0,0,0,5">
-                <TextBox Text="{Binding Host, UpdateSourceTrigger=PropertyChanged, ValidatesOnNotifyDataErrors=True}" x:Name="HostBox"/>
-                <TextBlock Text="Host (e.g. 192.168.1.20)" IsHitTestVisible="False" Foreground="Gray" Margin="5,0,0,0"
-                           VerticalAlignment="Center"
-                           Visibility="{Binding Text, ElementName=HostBox, Converter={StaticResource StringNullOrEmptyToVisibilityConverter}}"/>
-            </Grid>
-            <Grid Margin="0,0,0,5">
-                <TextBox Text="{Binding Port, UpdateSourceTrigger=PropertyChanged, ValidatesOnNotifyDataErrors=True}" x:Name="PortBox"/>
-                <TextBlock Text="Port" IsHitTestVisible="False" Foreground="Gray" Margin="5,0,0,0"
-                           VerticalAlignment="Center"
-                           Visibility="{Binding Text, ElementName=PortBox, Converter={StaticResource StringNullOrEmptyToVisibilityConverter}}"/>
-            </Grid>
-            <Grid Margin="0,0,0,5">
-                <TextBox Text="{Binding Username}" x:Name="UserBox"/>
-                <TextBlock Text="Username" IsHitTestVisible="False" Foreground="Gray" Margin="5,0,0,0"
-                           VerticalAlignment="Center"
-                           Visibility="{Binding Text, ElementName=UserBox, Converter={StaticResource StringNullOrEmptyToVisibilityConverter}}"/>
-            </Grid>
-            <PasswordBox PasswordChar="*" helpers:PasswordBoxAssistant.BindPassword="True" helpers:PasswordBoxAssistant.BoundPassword="{Binding Password}" Margin="0,0,0,5"/>
-            <StackPanel Orientation="Horizontal" Margin="0,0,0,5">
-                <Grid Width="180">
-                    <TextBox Text="{Binding LocalPath}" x:Name="LocalPathBox"/>
-                    <TextBlock Text="Local Path" IsHitTestVisible="False" Foreground="Gray" Margin="5,0,0,0"
-                               VerticalAlignment="Center"
-                               Visibility="{Binding Text, ElementName=LocalPathBox, Converter={StaticResource StringNullOrEmptyToVisibilityConverter}}"/>
-                </Grid>
-                <Button Content="Browse" Command="{Binding BrowseCommand}" Margin="5,0,0,0"/>
-            </StackPanel>
-            <Grid Margin="0,0,0,5">
-                <TextBox Text="{Binding RemotePath}" x:Name="RemotePathBox"/>
-                <TextBlock Text="Remote Path" IsHitTestVisible="False" Foreground="Gray" Margin="5,0,0,0"
-                           VerticalAlignment="Center"
-                           Visibility="{Binding Text, ElementName=RemotePathBox, Converter={StaticResource StringNullOrEmptyToVisibilityConverter}}"/>
-            </Grid>
-            <Button Content="Send" Command="{Binding TransferCommand}" Width="100"/>
+        <Grid.RowDefinitions>
+            <RowDefinition Height="Auto"/>
+            <RowDefinition Height="Auto"/>
+            <RowDefinition Height="*"/>
+            <RowDefinition Height="*"/>
+        </Grid.RowDefinitions>
+        <TextBlock Text="{Binding ServerStatus}" FontWeight="Bold" FontSize="14"/>
+        <StackPanel Grid.Row="1" Orientation="Horizontal" Margin="0,10,0,0">
+            <Button Content="Start" Width="100" Margin="0,0,5,0" Command="{Binding StartServerCommand}"/>
+            <Button Content="Stop" Width="100" Command="{Binding StopServerCommand}"/>
         </StackPanel>
-        <StackPanel Grid.Column="1" Margin="10">
-            <TextBlock Text="Logs" FontWeight="Bold"/>
-            <ComboBox x:Name="LogLevelBox" Width="120" Margin="0,0,0,5" SelectedIndex="0" SelectionChanged="LogLevelBox_SelectionChanged">
-                <ComboBoxItem Content="All"/>
-                <ComboBoxItem Content="Debug"/>
-                <ComboBoxItem Content="Warning"/>
-                <ComboBoxItem Content="Error"/>
-            </ComboBox>
-            <RichTextBox x:Name="LogBox" Height="200" IsReadOnly="True"/>
-        </StackPanel>
+        <GroupBox Header="Uploaded Files" Grid.Row="2" Margin="0,10,0,0">
+            <ListBox ItemsSource="{Binding UploadedFiles}">
+                <ListBox.ItemTemplate>
+                    <DataTemplate>
+                        <StackPanel Orientation="Horizontal">
+                            <TextBlock Text="{Binding Path}"/>
+                            <TextBlock Text=" ("/>
+                            <TextBlock Text="{Binding Size}"/>
+                            <TextBlock Text=" bytes)"/>
+                        </StackPanel>
+                    </DataTemplate>
+                </ListBox.ItemTemplate>
+            </ListBox>
+        </GroupBox>
+        <GroupBox Header="Downloaded Files" Grid.Row="3" Margin="0,10,0,0">
+            <ListBox ItemsSource="{Binding DownloadedFiles}">
+                <ListBox.ItemTemplate>
+                    <DataTemplate>
+                        <StackPanel Orientation="Horizontal">
+                            <TextBlock Text="{Binding Path}"/>
+                            <TextBlock Text=" ("/>
+                            <TextBlock Text="{Binding Size}"/>
+                            <TextBlock Text=" bytes)"/>
+                        </StackPanel>
+                    </DataTemplate>
+                </ListBox.ItemTemplate>
+            </ListBox>
+        </GroupBox>
     </Grid>
 </Page>
