--- DesktopApplicationTemplate.UI/Views/FtpServerView.xaml
@@ -1,43 +1,0 @@
-<Page x:Class="DesktopApplicationTemplate.UI.Views.FtpServerView"
-      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-      xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-      mc:Ignorable="d">
-    <Grid Margin="10">
-        <Grid.RowDefinitions>
-            <RowDefinition Height="Auto"/>
-            <RowDefinition Height="*"/>
-            <RowDefinition Height="*"/>
-        </Grid.RowDefinitions>
-        <TextBlock Text="{Binding ServerStatus}" FontWeight="Bold" FontSize="14"/>
-        <GroupBox Header="Uploaded Files" Grid.Row="1" Margin="0,10,0,0">
-            <ListBox ItemsSource="{Binding UploadedFiles}">
-                <ListBox.ItemTemplate>
-                    <DataTemplate>
-                        <StackPanel Orientation="Horizontal">
-                            <TextBlock Text="{Binding Path}"/>
-                            <TextBlock Text=" ("/>
-                            <TextBlock Text="{Binding Size}"/>
-                            <TextBlock Text=" bytes)"/>
-                        </StackPanel>
-                    </DataTemplate>
-                </ListBox.ItemTemplate>
-            </ListBox>
-        </GroupBox>
-        <GroupBox Header="Downloaded Files" Grid.Row="2" Margin="0,10,0,0">
-            <ListBox ItemsSource="{Binding DownloadedFiles}">
-                <ListBox.ItemTemplate>
-                    <DataTemplate>
-                        <StackPanel Orientation="Horizontal">
-                            <TextBlock Text="{Binding Path}"/>
-                            <TextBlock Text=" ("/>
-                            <TextBlock Text="{Binding Size}"/>
-                            <TextBlock Text=" bytes)"/>
-                        </StackPanel>
-                    </DataTemplate>
-                </ListBox.ItemTemplate>
-            </ListBox>
-        </GroupBox>
-    </Grid>
-</Page>
