--- DesktopApplicationTemplate.UI/Views/FtpServerView.xaml.cs
@@ -1,16 +1,0 @@
-using System.Windows.Controls;
-using DesktopApplicationTemplate.Core.Services;
-using DesktopApplicationTemplate.UI.ViewModels;
-
-namespace DesktopApplicationTemplate.UI.Views
-{
-    public partial class FtpServerView : Page
-    {
-        public FtpServerView(FtpServerViewViewModel viewModel, ILoggingService logger)
-        {
-            InitializeComponent();
-            DataContext = viewModel;
-            viewModel.Logger = logger;
-        }
-    }
-}
